Removed Add Points, Reset points button now also resets visited state.
